// src/business-center/practice-settings/adjustment-types/adjustment-types.ts
export class AdjustmentTypes {
    AdjustmentTypeId?: string;
    DataTag?: string;
    DateModified?: string;
    Description?: string;
    ImpactType?: number;
    IsActive?: boolean;
    IsAdjustmentTypeAssociatedWithTransactions?: boolean;
    IsDefaultTypeOnBenefitPlan?: boolean;
    IsPositive?: boolean;
    IsSystemType?: boolean;
    UserModified?: string;
}